Check all package source directories in current directory

check_dir() Identifies all R packages in the given directory (non-recursively) and passes them to the check_pkgs()

Usage

check_dir(
  path,
  n = 2L,
  output = tempfile(paste(utils::packageName(), Sys.Date(), sep = "-")),
  lib.loc = .libPaths(),
  repos = getOption("repos"),
  restore = options::opt("restore"),
  reporter = reporter_default(),
  ...
)

Arguments

path

file path to the package source directory

n

integer value indicating maximum number of subprocesses that can be simultaneously spawned when executing tasks.

output

character value specifying path where the output should be stored.

lib.loc

character vector with libraries allowed to be used when checking packages, defaults to entire .libPaths().

repos

character vector of repositories which will be used when generating task graph and later pulling dependencies.

restore

logical indicating whether output directory should be unlinked before running checks. If FALSE, an attempt will me made to restore previous progress from the same output (Defaults to NA, overwritable using option 'checked.restore' or environment variable 'R_CHECKED_RESTORE')

reporter

A reporter to provide progress updates. Will default to the most expressive command-line reporter given your terminal capabilities.

...

Additional arguments passed to checked-task-df and run()

Value

check_design() R6 class storing all the details regarding checks that run. Can be combined with results and summary() methods to generate results.

See also