Create a task to run R CMD check

Create a task to run R CMD check

Usage

check_task_spec(
  args = options::opt("check_args"),
  build_args = options::opt("check_build_args"),
  ...
)

Arguments

args

Character vector of arguments to pass to R CMD check. Pass each argument as a single element of this character vector (do not use spaces to delimit arguments like you would in the shell). For example, to skip running of examples and tests, use args = c("--no-examples", "--no-tests") and not args = "--no-examples --no-tests". (Note that instead of the --output option you should use the check_dir argument, because --output cannot deal with spaces and other special characters on Windows.)

build_args

Character vector of arguments to pass to R CMD build. Pass each argument as a single element of this character vector (do not use spaces to delimit arguments like you would in the shell). For example, build_args = c("--force", "--keep-empty-dirs") is a correct usage and build_args = "--force --keep-empty-dirs" is incorrect.

...

Arguments passed on to task_spec

alias

task alias which also serves as unique identifier of the task.

package_spec

package_spec object

env

environmental variables to be set in separate process running specific task.

See also