Check schedule data frame

Create data.frame which each row defines a package for which R CMD check should be run. Such data.frame is a prerequisite for generating check_design() which orchestrates all the processes including dependencies installation.

Arguments

path

path to the package source. Can be either a single source code directory or a directory containing multiple package source code directories.

Value

The check schedule data.frame with the following columns:

alias: The alias of the check to run. It also serves the purpose of providing a unique identifier and node name in the task graph.

version: Version of the package to be checked.

package: Object that inherits from check_task_spec(). Defines how package to be checked can be acquired.

custom: Object that inherits from custom_install_task_spec(). Defines custom package, for instance only available from local source, that should be installed before checking the package.

Details

_tasks_df() functions generate check task data.frame for all source packages specified by the path. Therefore it accepts it to be a vector of an arbitrary length.

See also