Check schedule data frame

Create data.frame which each row defines a package for which R CMD check should be run. Such data.frame is a prerequisite for generating check_design which orchestrates all the processes including dependencies installation.

rev_dep_check_tasks_df(
  path,
  repos = getOption("repos"),
  development_only = FALSE
)

source_check_tasks_df(path)

Arguments

path

path to the package source. See Details.

repos

repository used to identify reverse dependencies.

development_only

logical whether reverse dependency check should be run only against development version of the package. Applicable mostly when checking whether adding new package would break tests of packages already in the repository and taking the package as suggests dependency. Default to FALSE.

Value

The check schedule data.frame has strict structure and consists of following columns:

alias The alias of the check to run. It also serves the purpose of u unique identifier and node name in the task graph.

version Version of the package to be checked.

package Object that inherits from check_task_spec. Defines how package to be checked can be acquired.

custom Object that inherits from custom_install_task_spec. Defines custom package, for instance only available from local source, that should be installed before checking the package.

Details

rev_dep_check_tasks_df generates checks schedule data.frame appropriate for running reverse dependency check for certain source package. In such case path parameter should point to the source of the development version of the package and repos should be a repository for which reverse dependencies should be identified.

source_check_tasks_df generates checks schedule data.frame for all source packages specified by the path. Therefore it accepts it to be a vector of an arbitrary length.