Find the Next Packages Not Dependent on an Unavailable Package

While other packages are in progress, ensure that the next selected package already has its dependencies done.

task_graph_which_satisfied(
  g,
  v = igraph::V(g),
  dependencies = TRUE,
  status = STATUS$pending
)

task_graph_which_satisfied_strong(..., dependencies = "strong")

task_graph_which_check_satisfied(
  g,
  ...,
  dependencies = "all",
  status = STATUS$pending
)

task_graph_which_install_satisfied(
  g,
  ...,
  dependencies = "strong",
  status = STATUS$pending
)

Arguments

g

A dependency graph, as produced with task_graph_create().

v

Names or nodes objects of packages whose satisfiability should be checked.

dependencies

Which dependencies types should be met for a node to be considered satisfied.

status

status name. Nodes in v fill be filtered to consists only nodes with that status.

...

parametrs passed to down-stream functions.

Value

The name of the next package to prioritize

Details

There are helpers defined for particular use cases that strictly rely on the task_graph_which_satisfied, they are:

task_graph_which_satisfied_strong - List vertices whose strong dependencies are satisfied.

task_graph_which_check_satisfied - List root vertices whose all dependencies are satisfied.

task_graph_which_install_satisfied - List install vertices whose dependencies are all satisfied