checked Options

Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.

Checking Option Values

Option values specific to checked can be accessed by passing the package name to env.

options::opts(env = "checked")

options::opt(x, default, env = "checked")

Options

tty_tick_interval

tty refresh interval when reporting results in milliseconds
default:

0.1

option:

checked.tty_tick_interval

envvar:

R_CHECKED_TTY_TICK_INTERVAL (evaluated if possible, raw string otherwise)

results_error_on

character vector indicating whether R error should be thrown when issues are discovered when generating results. "never" means that no errors are thrown. If "issues" then errors are emitted only on issues, whereas "potential issues" stands for error on both issues and potential issues.
default:

"never"

option:

checked.results_error_on

envvar:

R_CHECKED_RESULTS_ERROR_ON (evaluated if possible, raw string otherwise)

results_keep

character vector indicating which packages should be included in the results. "all" means that all packages are kept. If "issues" then only packages with issues identified, whereas "potential_issues" stands for keeping packages with both "issues" and "potential_issues".
default:

"all"

option:

checked.results_keep

envvar:

R_CHECKED_RESULTS_KEEP (evaluated if possible, raw string otherwise)

restore

logical indicating whether output directory should be unlinked before running checks. If FALSE, an attempt will me made to restore previous progress from the same output
default:

NA

option:

checked.restore

envvar:

R_CHECKED_RESTORE (evaluated if possible, raw string otherwise)

check_envvars

named character vector of environment variables to use during R CMD check.
default:

c(`_R_CHECK_FORCE_SUGGESTS_` = FALSE, `_R_CHECK_RD_XREFS_` = FALSE,
    `_R_CHECK_SYSTEM_CLOCK_` = FALSE, `_R_CHECK_SUGGESTS_ONLY_` = TRUE)

option:

checked.check_envvars

envvar:

R_CHECKED_CHECK_ENVVARS (evaluated if possible, raw string otherwise)

check_build_args

character vector of args passed to the R CMD build.
default:

c("--no-build-vignettes", "--no-manual")

option:

checked.check_build_args

envvar:

R_CHECKED_CHECK_BUILD_ARGS (space-separated R CMD build flags)

check_args

character vector of args passed to the R CMD check.
default:

c("--timings", "--ignore-vignettes", "--no-manual")

option:

checked.check_args

envvar:

R_CHECKED_CHECK_ARGS (space-separated R CMD check flags)

See also