Plan Reverse Dependency Checks

Generates a plan for running reverse dependency check for certain source package. In such case path should be proivded with a directory path to the development version of the package and repos should be a repository for which reverse dependencies should be identified.

Usage

plan_rev_dep_checks(
  path,
  repos = getOption("repos"),
  remotes_dependencies = TRUE
)

Arguments

path

path to the package source.

repos

repository used to identify reverse dependencies.

remotes_dependencies

A vector of length one or a named list. Compatible with as_pkg_dependencies. Used to filter out remotes dependencies.

See also