Check results

Get R CMD check results

Usage

results(x, ...)

# S3 method for class 'check_design'
results(x, error_on = options::opt("results_error_on"), ...)

Arguments

x

check_design object.

...

other parameters.

error_on

character vector indicating whether R error should be thrown when issues are discovered when generating results. "never" means that no errors are thrown. If "issues" then errors are emitted only on issues, whereas "potential issues" stands for error on both issues and potential issues. (Defaults to "never", overwritable using option 'checked.results_error_on' or environment variable 'R_CHECKED_RESULTS_ERROR_ON')

See also