Build Tasks for Reverse Dependency Checks Generates checks schedule data.frame appropriate for running reverse dependency check for certain source package. In such case path parameter should point to the source of the development version of the package and repos should be a repository for which reverse dependencies should be identified.

Create data.frame which each row defines a package for which R CMD check should be run. Such data.frame is a prerequisite for generating check_design() which orchestrates all the processes including dependencies installation.

Usage

rev_dep_check_tasks_df(
  path,
  repos = getOption("repos"),
  versions = c("dev", "release"),
  lib.loc = .libPaths(),
  ...
)

Arguments

path

path to the package source. Can be either a single source code directory or a directory containing multiple package source code directories.

repos

repository used to identify reverse dependencies.

versions

character vector indicating against which versions of the package reverse dependency should be checked. c("dev", "release") (default) stands for the classical reverse dependency check. "dev" checks only against development version of the package which is applicable mostly when checking whether adding new package would break tests of packages already in the repository and take the package as suggests dependency.

lib.loc

vector of libraries used to check whether reverse dependency check can return accurate results.

...

parameters passed to the task specs allowing to customize subprocesses.

Value

The check schedule data.frame with the following columns:

alias: The alias of the check to run. It also serves the purpose of providing a unique identifier and node name in the task graph.

version: Version of the package to be checked.

package: Object that inherits from check_task_spec(). Defines how package to be checked can be acquired.

custom: Object that inherits from custom_install_task_spec(). Defines custom package, for instance only available from local source, that should be installed before checking the package.

Details

_tasks_df() functions generate check task data.frame for all source packages specified by the path. Therefore it accepts it to be a vector of an arbitrary length.

See also