Task specification

Create task specification list which consists of all the details required to run specific task.

task_spec(alias = NULL, package_spec = NULL, env = NULL)

install_task_spec(type = getOption("pkgType"), INSTALL_opts = NULL, ...)

custom_install_task_spec(...)

check_task_spec(args = NULL, build_args = NULL, ...)

revdep_check_task_spec(revdep, ...)

Arguments

alias

task alias which also serves as unique identifier of the task.

package_spec

package_spec object

env

environmental variables to be set in separate process running specific task.

type

character, indicating the type of package to download and install. Will be "source" except on Windows and some macOS builds: see the section on ‘Binary packages’ for those.

INSTALL_opts

an optional character vector of additional option(s) to be passed to R CMD INSTALL for a source package install. E.g., c("--html", "--no-multiarch", "--no-test-load").

Can also be a named list of character vectors to be used as additional options, with names the respective package names.

...

parameters passed to downstream constructors

args

Character vector of arguments to pass to R CMD check. Pass each argument as a single element of this character vector (do not use spaces to delimit arguments like you would in the shell). For example, to skip running of examples and tests, use args = c("--no-examples", "--no-tests") and not args = "--no-examples --no-tests". (Note that instead of the --output option you should use the check_dir argument, because --output cannot deal with spaces and other special characters on Windows.)

build_args

Character vector of arguments to pass to R CMD build. Pass each argument as a single element of this character vector (do not use spaces to delimit arguments like you would in the shell). For example, build_args = c("--force", "--keep-empty-dirs") is a correct usage and build_args = "--force --keep-empty-dirs" is incorrect.

revdep

character indicating whether the task specification describes check associated with the development (new) or release (old) version of the for which reverse dependency check is run.